Added logic for saved location

+++ app/src/main/java/com/example/wesniemarcelin/chaseweatherappcodingchallenge/model/Wind.java
@@ -6,7 +6,7 @@ package com.example.wesniemarcelin.chaseweatherappcodingchallenge.model;
 
 public class Wind {
     private float speed;
-    private int deg;
+    private float deg;
 
     public float getSpeed() {
         return speed;
@@ -16,7 +16,7 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
